Fix passing file GLOBs to $c->res->body
Contrary to the documentation of Catalyst::Response, passing a
filehandle GLOB did not work, since the content-length header was not
computed correctly.

body checked if it was passed a blessed object and only then use stat
to get the file size. In all other cases the passed thing was considered
a string and length was used. Filehandle GLOBs are neither blessed nor
strings, so the content-length header always got set to 15 (length of a
stringified GLOB: "GLOB(0x123456)").