Stop psgi_file.t from recursing
psgi_app is *not* to be used in .psgi files. It's the compat layer that'll give
you either an app loaded from from a .psgi file, or the default PSGI app with
the reverse proxy middleware and stuff.

In .psgi files only raw_psgi_app should ever be used, as that'll return the bare
catalyst psgi app to then be wrapped in middlewares manually.

I see how the naming is shit and confusing though. Suggestions welcome.