Make getting to a PSGI app saner
All the user should ever need to call is ->psgi_app (previously raw_psgi_app),
to get an unwrapped psgi app as provided by the engine. This is what should be
used when writing .psgi files.

For all other ways of running catalyst apps there's no need to be able to get to
the psgi application that'll actually run, so here we're making everything
that's responsible for reading .psgi files to run them with Catalyst::Script::*
or for wrapping the plain psgi code ref with back-compat middleware private.

Unfortunately this removes some of the docs about the automatic .psgi
loading. These should be added again elsewhere. Also the Upgrading docs need
updating after this.