No, we don't want this
If we'd check it, we'd have to check it on every request. That's entirely not
worth it as I'm not even aware of any server that doesn't support psgi.streaming
right now, let alone a server people actually use.

If people complain, documentation suggesting to load the BufferedWrite
middleware needs to be written.