Fix model/view/controller methods to handle stringifiable objects
Previously, we tested ref() and assumed if that was true that we'd been
supplied with a regexp; by switching to using Safe::Isa to check for

  $thing->$_isa('Regexp')

a stringifiable object will still get handled correctly. This is necessary
when using string overloading tools such as i18n.pm or HTML::String.