Unicode plugin - rework encoding default config
The aim is to find a way to: 1) not bust people who are doing the right
thing, 2) force people to do the right thing going forward, with new
apps, 3) make it easy for people doing the wrong thing to keep doing
the wrong thing if they don't or can't fix their apps.

Thus, lets basically set "encoding" bit to 'UTF-8' for apps loading the
plugin directly (unless provided something else) and remove the default
value.