Unicode plugin - rework exception handler
For apps in the third case (doing its own decoding/encoding), we are
not setting an ->encoding object and then request cycle might blow
up during prepare time (when we would do the params decoding). Thus,
we change the default behavior from rethrow of this exception to a
simple warning.