Added 'autoflush' to Catalyst::Log + enabled during setup
The default logger now has an 'autoflush' flag to cause log messages to
be written immediately instead of queued. This is now set to true during
setup so that log messages show up in real time (i.e. when using the
test server).

It is currently turned back off at the end of setup for backward
compatability (specifically for Static::Simple), however, mst has
suggested that we change the API such that plugins like S::S are
responsible for turning off autoflush themselves if they need to.

See in-code comments for more details.