First proof of concept for  Catalyst::Utils::filehandle_response_at_path
Actually, I begin to think this is not so much useful. It comes to a price of
additional dependencies. Then there should be a way to override the mimetypes
(for example, the epub in the test file has unwanted side-effects, and I would
need it to be application/epub+zip).

So we end up with another piece or either overly complicated code, or pretty
useless code.

The programmer can just stuff the IO::File::WithPath into the response
with recent catalyst and set headers as needed. But this change is
for the cutting edge catalyst, so...