Travis: preliminary reverse dependency testing
The travis config will now run the test suites for some of the more
well-known reverse dependencies (i.e. for which Catalyst is a
dependency) after the main test suite is ran. This is meant to catch
potential breakage of these downstream packages for which Catalyst
should maintain proper support. These include standard plugins such as
Catalyst::Plugin::Authentication and Catalyst::Plugin::Static::Simple.
Modules in Task::Catalyst were used to compile this list, which is still
preliminary. We should add more as we go (however, we do need to be
selective to keep the travis run time below 50 mins at which point it
will timeout)

Note:

As of this commit and after the 5.90070 release, we already expect the
reverse deps for the following to fail:

 Catalyst-Plugin-Static-Simple-0.32
 Catalyst-Plugin-Session-0.39

Catalyst-Plugin-Authentication-0.10023 was also failing as of 5.90070,
but has already been fixed in 52af51596