stop using cached Hash::MultiValue from env
Plack::Request and Catalyst::Request have incompatible semantics for
naked parameters: P::Rq parses them as keys with empty-string values
while C::Rq parses them as keys with an undef value. Therefore even
if a Hash::MultiValue object was cached by P::Rq in the environment,
it is wrong to reuse it.