Fix handling of empty paths in uri_for
In versions of Catalyst before 5.90097, it was possible to call `uri_for`
with an empty path and receive the URI of the relevant action (as mentioned
in the documentation for `uri_for`).  From version 5.90097, this caused the
following warning (with large backtrace) to be thrown:

    Use of uninitialized value $path in substitution (s///) at
    $HOME/perl5/perlbrew/perls/perl-5.18.4/lib/site_perl/5.18.4/Catalyst.pm
    line 1564.

This was because even though the `$path` wasn't blessed, it hadn't been
initialised to any value and thus was throwing a warning on the substitution
operation.

This change reinstates the previous behaviour (which is still documented in
current versions) and adds a test to protect against future regressions.