Revert incorrect fixes for parsing of attributes with empty parens
This partially reverts commits:

  • "Args() wasn't being processed as unlimited number of args, due to…" (43b44b3)
  • "Fixed :Path() empty brackets attribute not registering action. Fixes: #104" (dd6a9f2)

leaving the tests added by them for the Path() bug in GH #104 but
reverting the code changes.

The code changes caused Attribute() to parse equivalent to Attribute('')
instead of Attribute.  The mistake was not noticed because Path, Path(),
and Path('') should all produce equivalent behaviour.  However, the same
is not true of other attributes.  For example, Args.

Parsing empty parens as undef rather than an empty string is the correct
solution.  Empty strings are achieved explicitly with Attribute('').

Reverting these changes fixes the regression in the handling of Args(),
which started producing the following error at app startup under debug
mode:

    Out of memory during list extend
    at .../Catalyst/DispatchType/Chained.pm line 101.

The following commit will make the correct fix.