Test for recent regression parsing Args()
Args() started parsing as Args('') instead of Args, which in
DispatchType::Chained->list's debugging output ends up causing the
equivalent of this to run:

    ("*") x ~0

The result is Perl throwing an "Out of memory during list extend" error.
Fix to follow.

NB: Using the normalized_arg_number (where ~0 stands in for ∞) for this
debugging output is dubious, but there's not a clear cut revert as the
changes which got us there are spread across multiple commits related to
the new typed Args support.  I'm saving any change to the use of
normalized_arg_number and list_extra_info for another time (and probably
another person), in favor of changing the least while fixing both the
Args() regression and the original Path() bug which precipitated it.