Change behaviour for terminal width detection
Sometimes Term::Size::Any is installed but can return a width of
undef. This causes the eval to succeed but then it falls back to
using 80 characters.

This change allows it to look at $ENV{COLUMNS} for a hint if the
width isn't defined or if the eval fails due to a missing module
instead of just if the eval returns an error.