Avoid List::MoreUtils
The latest version of this module has a confusing and hard-to-honour set of
licences, with different terms for code added in different versions.

The only use we make of it is the uniq() function. A routine with the same
behaviour is available in newer versions of List::Util, and we already
depend on older versions of *that* module. So depending on a recent enough
version of List::Util means that this change actually reduces the number of
non-core dependencies for users with a new enough version of Perl.