prevent -h being an alias for --help, since we use it for --host
The -h flag is meant to be an alias for --host, but MooseX::Getopt added
it as an alias for --help in version 0.46 (2012-08-30). This conflict
has been silently ignored since then, until Getopt::Long::Descriptive
added a warning for this in version 0.106 (2021-03-12).

Modify the help option to remove the -h alias. We could try to filter
the 'h' option out, but it's easier to just hard code the aliases we
want to allow.