fix handling of slurpy arguments in tuples
Type::Tiny version 1.016002 changed how slurpy types are represented.
Catalyst had been looking at the original parameters to a Tuple type,
and expecting a slurpy type to be a hashref as the last parameter. That
is no longer true, so the argument count checking would fail.

For new Type::Tiny versions, we can check that the final parameter is a
Slurpy type. This seems slightly more robust.