Clearer attribute unquoting
Replace the two regexes with one.

This also fixes a subtle already-present bug in the double-quote
unquoting where it's missing the anchor to the end of string - I think
that's not actually bitten anyone because the `.*` capture is greedy, so
it'll capture as much as it can, so the `"` at the end will match the
last double quote, not the first one it sees.